third_party/blink/web_tests/platform/linux-chrome/external/wpt/webrtc/RTCPeerConnection-createDataChannel-expected.txt

This is a testharness.js-based test.
Found 4 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] createDataChannel attribute default values
  assert_equals: expected "blob" but got "arraybuffer"
[FAIL] createDataChannel with provided parameters should initialize attributes to provided values
  assert_equals: expected "blob" but got "arraybuffer"
[FAIL] createDataChannel with id 65535 and negotiated not set should succeed, but not set the channel's id
  Failed to execute 'createDataChannel' on 'RTCPeerConnection': RTCDataChannel cannot have id > 65534
[FAIL] createDataChannel with id 65534 and negotiated true should succeed, and set the channel's id
  Failed to execute 'createDataChannel' on 'RTCPeerConnection': RTCDataChannel creation failed
Harness: the test ran to completion.
